# docs/openapi.yaml
components:
  schemas:
...
    DatabaseItemPageResponse:
      description: Paginated food database response.
      properties:
        items:
          items:
            $ref: '#/components/schemas/DatabaseItemResponse'
          title: Items
          type: array
        limit:
          title: Limit
          type: integer
        offset:
          title: Offset
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - limit
      - offset
      title: DatabaseItemPageResponse
      type: object
    DatabaseItemResponse:
      description: Unified food database row for desktop database management.
      properties:
        aliases:
          items:
            type: string
          title: Aliases
          type: array
        carbs_g:
          anyOf:
          - type: number
          - type: 'null'
          title: Carbs G
        default_grams:
          anyOf:
          - type: number
          - type: 'null'
          title: Default Grams
        display_name:
          title: Display Name
          type: string
        fat_g:
          anyOf:
          - type: number
          - type: 'null'
          title: Fat G
        fiber_g:
          anyOf:
          - type: number
          - type: 'null'
          title: Fiber G
        id:
          format: uuid
          title: Id
          type: string
        image_cache_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Cache Path
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        is_verified:
          default: false
          title: Is Verified
          type: boolean
        kcal:
          anyOf:
          - type: number
          - type: 'null'
          title: Kcal
        key:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
        kind:
          enum:
          - pattern
          - product
          - restaurant
          title: Kind
          type: string
        last_used_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Used At
        nutrients_json:
          additionalProperties: true
          title: Nutrients Json
          type: object
        prefix:
          anyOf:
          - type: string
          - type: 'null'
          title: Prefix
        protein_g:
          anyOf:
          - type: number
          - type: 'null'
          title: Protein G
        quality_warnings:
          items:
            type: string
          title: Quality Warnings
          type: array
        source_confidence:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Confidence
        source_file:
          anyOf:
          - type: string
          - type: 'null'
          title: Source File
        source_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Name
        source_page:
          anyOf:
          - type: integer
          - type: 'null'
          title: Source Page
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        subtitle:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtitle
        token:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
        usage_count:
          default: 0
          title: Usage Count
          type: integer
      required:
      - id
      - kind
      - display_name
      title: DatabaseItemResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  summary: Informational food diary API.
  title: Glucotracker API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /database/items:
    get:
      description: List local pattern and product database rows for desktop management.
      operationId: listDatabaseItems
      parameters:
      - in: query
        name: type
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
      - in: query
        name: source
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
      - in: query
        name: q
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - in: query
        name: needs_review
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Needs Review
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseItemPageResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: List Database Items
      tags:
      - database
      x-glucotracker-scoped: true
...
  /photos/{photo_id}/file:
    get:
      description: Stream the stored image bytes for a photo.
      operationId: getPhotoFile
      parameters:
      - in: path
        name: photo_id
        required: true
        schema:
          format: uuid
          title: Photo Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Photo File
      tags:
      - photos
      x-glucotracker-scoped: true
...